requirements.txt
pandas
scikit-learn
gradio
python-dotenv
papermill
pathlib